packages/backend/pyproject.toml
[project]
name = "tasche"
version = "0.1.0"
description = "Tasche Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "python-ulid>=2.0.0",
    "mangum>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tasche"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
]
